+++ test-packages/test-services/v2/test-service/TestEntityLvl2MultiLinkApi.ts
@@ -20,8 +20,6 @@ import {
   EdmTypeField,
   OrderableEdmTypeField
 } from '@sap-cloud-sdk/odata-v2';
-import { BigNumber } from 'bignumber.js';
-import { Moment } from 'moment';
 export class TestEntityLvl2MultiLinkApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements EntityApi<TestEntityLvl2MultiLink<DeSerializersT>, DeSerializersT>