+++ test-packages/test-services/v2/test-service/TestEntitySingleLinkApi.ts
@@ -26,8 +26,6 @@ import {
   Link,
   OneToOneLink
 } from '@sap-cloud-sdk/odata-v2';
-import { BigNumber } from 'bignumber.js';
-import { Moment } from 'moment';
 export class TestEntitySingleLinkApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements EntityApi<TestEntitySingleLink<DeSerializersT>, DeSerializersT>