+++ test-packages/test-services/v4/test-service/TestEntityApi.ts
@@ -32,8 +32,6 @@ import {
   OneToManyLink,
   OneToOneLink
 } from '@sap-cloud-sdk/odata-v4';
-import { BigNumber } from 'bignumber.js';
-import { Moment, Duration } from 'moment';
 export class TestEntityApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements EntityApi<TestEntity<DeSerializersT>, DeSerializersT>