+++ test-packages/test-services/v4/test-service/TestEntityLvl2MultiLinkApi.ts
@@ -23,8 +23,6 @@ import {
   OrderableEdmTypeField,
   OneToManyLink
 } from '@sap-cloud-sdk/odata-v4';
-import { BigNumber } from 'bignumber.js';
-import { Moment, Duration } from 'moment';
 export class TestEntityLvl2MultiLinkApi<
   DeSerializersT extends DeSerializers = DefaultDeSerializers
 > implements EntityApi<TestEntityLvl2MultiLink<DeSerializersT>, DeSerializersT>